Validação 'void' - Finalização em andamento

+++ projeto_sqlite/crud.py
@@ -1,10 +1,10 @@
 import sqlite3
 from prettytable import PrettyTable
-
+import os
 
 # ========================== COMANDO EXIBIR =============
 class Exibir:
-    def exibir(self, tabela):
+    def SELECT(self, tabela):
         conn = sqlite3.connect("C:/vinicius/sqlite/projeto_sqlite/aeroporto.db")
         cursor = conn.cursor()
 
@@ -28,15 +28,35 @@ class Exibir:
 
 # ========================== COMANDO ADICIONAR =============
 class Adicionar:
-    def adicionar(self, tabela):
+    def INSERT(self, tabela):
+        conn = sqlite3.connect("C:/vinicius/sqlite/projeto_sqlite/aeroporto.db")
+        cursor = conn.cursor()
+        
+        # verifica se a tabela não esta vazia ou se é uma instancia invalida
+        if not tabela or not isinstance(tabela, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
         conn = sqlite3.connect("C:/vinicius/sqlite/projeto_sqlite/aeroporto.db")
         cursor = conn.cursor()
 
-        print()
+
         colunas = input('Digite as colunas (separados por vírgula): ').split(',')
         dados = input('Digite os dados (separados por vírgula): ').split(',')
+        
+        # Verifica se as colunas ou dados estão vazios
+        if not colunas or not dados or len(colunas) != len(dados):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
 
-       # o 1° join cria uma lista com as colunas 2°join fazuma lista com os placeholders '?', que os valores que vao ser inseridos
         cursor.execute(f"INSERT INTO {tabela} ({', '.join(colunas)}) VALUES ({', '.join(['?'] * len(colunas))})", dados)
         
         conn.commit()
@@ -44,19 +64,44 @@ class Adicionar:
 
 # ========================== COMANDO ATUALIZAR =============
 class Atualizar:
-    def atualizar(self, tabela, identificador, valor):
+    def UPDATE(self, tabela, identificador, valor):
         conn = sqlite3.connect("C:/vinicius/sqlite/projeto_sqlite/aeroporto.db")
         cursor = conn.cursor()
-
-        print()
-        colunas = input('Digite as colunas que seram atualizadas (separados por vírgula): ').split(',')
-        dados = input('Digite os novos dados (separados por vírgula): ').split(',')
-
         
+        # Verificação se são válidos e se não estao vazios
+        if not tabela or not isinstance(tabela, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
+        if not identificador or not isinstance(identificador, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
+        if not valor or not isinstance(valor, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
+        colunas = input('Digite as colunas que serão atualizadas (separadas por vírgulas): ').split(',')
+        dados = input('Digite os novos dados (separadas por vírgulas): ').split(',')
+
         if len(colunas) != len(dados):
             print('+----------------------------------------+')
             print("+   O número de colunas não corresponde  +")
             print("+      ao número de dados fornecidos.    +")
+            print('+----------------------------------------+')
+            input('Pressione ENTER para voltar')
             return
 
         # Atualiza os registros
@@ -65,19 +110,42 @@ class Atualizar:
             dado = dados[i].strip()  
             cursor.execute(f"UPDATE {tabela} SET {coluna} = ? WHERE {identificador} = ?", (dado, valor))
 
-        
         conn.commit()
         conn.close()
 
 
 # ========================== COMANDO APAGAR =============
 class Apagar:
-    def apagar(self, tabela, identificador, valor):
+    def DELETE(self, tabela, identificador, valor):
         conn = sqlite3.connect("C:/vinicius/sqlite/projeto_sqlite/aeroporto.db")
         cursor = conn.cursor()
-
-       
-        cursor.execute(f"DELETE FROM {tabela} WHERE {identificador} = ?", (valor,))
         
+        # Verificação se são válidos e se não estao vazios
+        if not tabela or not isinstance(tabela, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
+        if not identificador or not isinstance(identificador, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
+        if not valor or not isinstance(valor, str):
+            os.system('cls') 
+            print('+------------------------------+')
+            print('Opção inválida. Tente novamente.')
+            print('+------------------------------+') 
+            input('Pressione ENTER para voltar')
+            return
+
+        cursor.execute(f"DELETE FROM {tabela} WHERE {identificador} = ?", (valor,))
+
         conn.commit()
         conn.close()
